+++ packages/ssg/src/lib/site/__fixtures__/site/site.config.ts
@@ -1,6 +1,0 @@
-import { defineSite } from '../../../../site.ts'
-import { markdown } from './transforms/markdown.tsx'
-
-export default defineSite(({ base, islandUrls }) => ({
-  transforms: [markdown({ base, islandUrls })],
-}))